refactor(scanner): extract strategy/regime.py — detect_regime + per-symbol + helpers (#225 PR6)
Moves to strategy/regime.py:
- detect_regime, get_cached_regime, detect_regime_for_symbol
- 5x _compute_*_score helpers (price, fng, funding, rsi, adx)
- _regime_cache_key, _compute_local_regime
- _load_regime_cache, _save_regime_cache
- _REGIME_CACHE_FILE/PATH, _REGIME_TTL_SEC constants
- _regime_cache module-global dict (identity preserved via re-export)

Critical fixture update: scanner_frozen.py monkeypatch targets switched from
btc_scanner.* to strategy.regime.* (home module). Also fixes same namespace
issue in test_regime_per_symbol.py and test_snapshot_regression.py.

btc_scanner.py: 1190 → 814 LOC (-376). All 16 names re-exported for backward compat.
